+++ app/srcs/headers/window_utils.hpp
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/13 13:53:15 by craimond          #+#    #+#             */
-/*   Updated: 2024/06/14 00:37:47 by craimond         ###   ########.fr       */
+/*   Updated: 2024/06/14 12:54:06 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@
 void	init_window(sf::RenderWindow &window);
 void	put_tile_on_window(sf::RenderWindow &window, const Tile &tile);
 void	put_grid_on_window(sf::RenderWindow &window, const Grid &grid);
-bool	is_mouse_in_window(const sf::RenderWindow &window, const Vector2D<int32_t> &mouse_pos);
+bool	is_mouse_in_window(const sf::RenderWindow &window, const Vector2D<int16_t> &mouse_pos);
 
 extern bool	visualization_in_progress;
